deps/v8/test/inspector/runtime/console-context-expected.txt

Tests console.context

Running test: testConsoleContextMethod
console.context description:
{
    className : Function
    description : function context() { [native code] }
    objectId : <objectId>
    type : function
}
console.context() methods:
[
    [0] : debug
    [1] : error
    [2] : info
    [3] : log
    [4] : warn
    [5] : dir
    [6] : dirXml
    [7] : table
    [8] : trace
    [9] : group
    [10] : groupCollapsed
    [11] : groupEnd
    [12] : clear
    [13] : count
    [14] : assert
    [15] : markTimeline
    [16] : profile
    [17] : profileEnd
    [18] : timeline
    [19] : timelineEnd
    [20] : time
    [21] : timeEnd
    [22] : timeStamp
]

Running test: testDefaultConsoleContext
undefined
undefined
undefined

Running test: testAnonymousConsoleContext
anonymous#2
anonymous#3
anonymous#4

Running test: testNamedConsoleContext
named-context#6
named-context#6
named-context#6

Running test: testTwoConsoleContextsWithTheSameName
named-context#7
named-context#8

Running test: testConsoleCountInDifferentConsoleContexts
[
    [0] : {
        type : string
        value : 239: 1
    }
]
[
    [0] : {
        type : string
        value : 239: 1
    }
]

Running test: testConsoleCountForNamedConsoleContext
[
    [0] : {
        type : string
        value : 239: 1
    }
]
[
    [0] : {
        type : string
        value : 239: 2
    }
]
